+++ CHANGELOG.md
@@ -5,6 +5,14 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Changed
+ - Refactor command palette props
+
+### Fixed
+ - Unable to close Command palette through editor focus
+ - #59 (Command palette selected keyboard and mouse)
+
+## [0.5.0] - 2017-12-12
 ### Removed
  - CDN Roboto
  - Old editor props
@@ -173,6 +181,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
  - Live html rendered preview column
  - CodeMirror editor
 
+[0.5.0]: https://github.com/grissius/markup-editor/compare/v0.4.0...v0.5.0
 [0.4.0]: https://github.com/grissius/markup-editor/compare/v0.3.1...v0.4.0
 [0.3.1]: https://github.com/grissius/markup-editor/compare/v0.3.0...v0.3.1
 [0.3.0]: https://github.com/grissius/markup-editor/compare/v0.2.0...v0.3.0